Consolidate ReleaseCertified funcs
