+++ cnf-certification-test/certification/certtool/certtool.go
@@ -17,14 +17,11 @@
 package certtool
 
 import (
-	"strings"
 	"time"
 
-	version "github.com/hashicorp/go-version"
 	"github.com/test-network-function/cnf-certification-test/internal/api"
 	"github.com/test-network-function/cnf-certification-test/pkg/configuration"
 	"github.com/test-network-function/cnf-certification-test/pkg/provider"
-	"helm.sh/helm/v3/pkg/release"
 )
 
 const (
@@ -67,44 +64,6 @@ func WaitForCertificationRequestToSuccess(certificationRequestFunc func() (inter
 		elapsed += pollingPeriod
 	}
 	return result
-}
-
-// CompareVersion compare between versions
-func CompareVersion(ver1, ver2 string) bool {
-	ourKubeVersion, _ := version.NewVersion(ver1)
-	kubeVersion := strings.ReplaceAll(ver2, " ", "")[2:]
-	if strings.Contains(kubeVersion, "<") {
-		kubever := strings.Split(kubeVersion, "<")
-		minVersion, _ := version.NewVersion(kubever[0])
-		maxVersion, _ := version.NewVersion(kubever[1])
-		if ourKubeVersion.GreaterThanOrEqual(minVersion) && ourKubeVersion.LessThan(maxVersion) {
-			return true
-		}
-	} else {
-		kubever := strings.Split(kubeVersion, "-")
-		minVersion, _ := version.NewVersion(kubever[0])
-		if ourKubeVersion.GreaterThanOrEqual(minVersion) {
-			return true
-		}
-	}
-	return false
-}
-
-func IsReleaseCertified(helm *release.Release, ourKubeVersion string, out api.ChartStruct) bool {
-	for _, entryList := range out.Entries {
-		for _, entry := range entryList {
-			if entry.Name == helm.Chart.Metadata.Name && entry.Version == helm.Chart.Metadata.Version {
-				if entry.KubeVersion != "" {
-					if CompareVersion(ourKubeVersion, entry.KubeVersion) {
-						return true
-					}
-				} else {
-					return true
-				}
-			}
-		}
-	}
-	return false
 }
 
 func GetContainersToQuery(env *provider.TestEnvironment) map[configuration.ContainerImageIdentifier]bool {
